ebenezer/requirements/local.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.1.2
django-jet==1.0.7
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.8.2
idna==2.7
numpy==1.15.2
pytz==2018.5
requests==2.19.1
urllib3==1.23
